requirements.txt
numpy>=1.12
Pillow